+++ Assets/Scripts/Scenes/Game/UI/EventData/EventData02.cs
@@ -20,7 +20,7 @@ namespace Akimichi.Game
             base.OnFinished();
             if (this.resultType != EventConst.EventMessageType.Yes) return;
             var send = DataObjectManager.Instance().Get();
-            send.Datas[0] = (int)PlayerManager.Instance().PlayerIndex;
+            send.Datas[0] = (byte)PlayerManager.Instance().PlayerIndex;
             send.Datas[1] = this.value;
             NetworkManager.Instance().SendEvent(EventConst.Event.AddWeight, send);
         }
